SI-5739 (bis) vals for subpatterns unless -optimize
To facilitate debugging pattern matches, we store the values for
sub-patterns of extractor (synthetic or user-defined) patterns in local variables.
When performing an optimized build, and when possible, we don't do store but inline them directly.

For soundness, SI-5158, SI-6070, we must always store the values of mutable case class fields.

(Specifying -optimize is the only way to suppress emitting these local variables.
An unoptimized build will always generate them, which was deemed the right default during the meeting.)

(updated flags for t4425 to get consistent runs on optimized and non-optimized partest runs
 by always passing -optimize)