Objects are now consistently called objects, not modules.
The good thing is that even if we choose a different name,
this commit serves as a more or less precise documentation
of places we need to touch to fix the API.

Depending on the amount of external projects which use these APIs,
it would probably make sense to add the old names back as
@deprecated members.

MODULE$ and other significant binary compatibility artifacts
were left intact.

Review by @paulp, @xeno-by, @adriaanm (is the isObject -> isInstance
rename in Patterns acceptable or is there a better name?) and
everyone else interested.