Moved five methods from LinearSeqOptimized to List
Curiously, profiling showed that these methods (in particular map) got slower when specialized in LinearSeqOptimized from their generic impl in TraversableLike.
I speculate that the virtual calls to head/tail/nonEmpty are more expensive than the closure creation which we save (Note that foreach IS specialized in List,
so head/tail are monomorphic, even though isEmpty still has two implementations.

If we confirm a speedup in this commit it would make sense to specialize more methods in LinearSeqOptimized again in List.