Added specialized implementation for filter
I had to use null instead of HashSet.empty for the empty set inside filter0 to get decent performance in the case where the filter predicate is almost always false. Validated via Google Caliper.

The new implementation has significant performance advantages. But the main point is that the HashSet1 objects of the original set are reused in the filtered set, which means that a) there are less object allocations and b) the hash code does not have to be recalculated for each element matching the filter predicate. For complex objects this can be a big performance win.