SI-6280 Scaladoc: Reloading preserves anchors
Scaladoc redirects from */package.html pages to index.html with the
qualified package name as the anchor. This is done to pop up the left
navigation panel whenever someone links directly to the page. (so
users may navigate the rest of the scaladoc pages)

When linking members directly, one can use */package.html#<signature>,
which translates to index.html#*.package@<signature>. The problem was
that the redirection javascript did not take the <signature> into
account, so it would redirect to the top of the package page instead
of scrolling to the member.

Note: this is a layout change, so there's no test for it.