rehash of source file information for Symbol
Symbol.pos is moved to macros.Universe, because it's empty at runtime.

Also added Symbol.associatedFile, because it has valid usages during runtime.
When reflecting one might want to get to a classfile of the symbol and then
analyze it to e.g. decompile the bytecode or find out the associated src file.