Exposes Symbol.sourceModule
As discussed, this is not the best API to expose, because this is an
implementation detail that might change.

However the ship has sailed. We're already imposing
the moduleClass <-> sourceModule quirk to the users of our API:
Evidence: http://stackoverflow.com/questions/12128783.

There are reflection tasks that cannot be pulled without the knowledge
of this implementation detail, so we shouldn't pretend that we can
change it on a whim and not break anything.

Hence I propose to add sourceModule to the public contract
and bear the potential consequences.