Reflection Sealing
This pull request seals the isXXX and asXXX methods in the traits
defined in the cake piece `scala.reflect.base.Symbols`. All isXXX
and asXXX methods which are known at a particular level to return
`false` or to throw an exception respectively (as defined in the
base abstraction) are overridden with a final method with the same
behavior as the overridden one.

Two of these methods cannot be overridden as this leads to a build
error (see https://issues.scala-lang.org/browse/SI-6298).