Made instrumenter more robust by looking at tokens
Made instrumenter more robust by taking into account the positions of the tokens in the source. This allows us to reliably put instrumentation code at the end of the previous token, which tends to put it at end of lines. Furthermore, we now skip left parents. Previously, the instrumenter got confused for a statement like

( x + 1 )

because it thought that the statement started at the `x` and not at the `(`.

Another small change is that we now use decoded names in the worksheet. So ??? will show as ??? not $qmark$qmark$qmark.
(cherry picked from commit b2ecb931e13abb18c25ac6d49c18e974e1dd793e)

Conflicts:

	src/compiler/scala/tools/nsc/interactive/ScratchPadMaker.scala