SI-6274 Fix owners when eta-expanding function with byName param
When eta-expanding a function that takes a by-name param the local definition
for the corresponding argument is a function-0

  val eta$1 = () => { argument-to-by-name }

If there are any definitinos in the `argument-to-by-name`, the symbol
owner needs to be changed to the anonymous function's symbol.

To know the function symbol in eta expand, we need to type-check the
function, and therefore pass the `Typer` instance to `etaExpand`.