SI-6323 prohibits reflection against free types
One of the use cases for free types is reification of local classes.

The result is very seamless. Despite that local classes are not pickled,
free types recreated the symbols referenced by local classes, so that we
get our own symbol table, which can be analyzed with usual precision of
pickled symbols and types.

However when we try to use those symbols for reflection, we hit a problem.
Scala runtime reflection uses its own mechanism for dealing with non-pickled
types, which is incompatible with mini symbol tables used in free types.

Therefore to prevent confusion, I prohibit using those symbols for reflection.