Fixes SI-6268. Review by @paulp and @lrytz.
This reverts a refactoring from https://github.com/scala/scala/commit/963aabbeb4

MultiString settings would not properly write an unparse string that could be
reparsed, leading to failures when forking scalac in ant. Specifically, if a
setting was empty, it was getting added to the unparse string and causing
scalac to fail. This at least reverts to previous behavior (also more correct
for multiple values). Whatever we do here has to work with the @file style
argument reading and can't place empty options on the command line.

Also, This assumes there are tests around the behvior the REPL needs.

Note: The test is not automated yet, so unfortunately, we need to improve ANT testing capabilities to prevent regressons