Rescued TreeBuilder from the parser.
For too long, Tree Builder has ruled over those who would
like to build their own trees.  Today marks the start of a
new era.  Trees are for building, not for parsers to hoard.
It's in Global now.

The particular motivation is exposing makeNew, and I also
added makeAnonymousNew, so rather than this...

Apply(Select(
  Block(List(ClassDef(
    Modifiers(scala.tools.nsc.symtab.Flags.FINAL), tpnme.ANON_CLASS_NAME, Nil,
    Template(List(Ident(definitions.AnyRefClass)), emptyValDef, List(
      DefDef(NoMods, nme.CONSTRUCTOR, Nil, List(Nil), TypeTree(), Block(
        List(Apply(Select(Super(This(tpnme.EMPTY), tpnme.EMPTY), nme.CONSTRUCTOR), Nil)),
        Literal(Constant(()))
      )))))),
  Apply(Select(New(Ident(tpnme.ANON_CLASS_NAME)), nme.CONSTRUCTOR), Nil))
, sn.GetClass), Nil)

We can write this.

  Apply(Select(makeAnonymousNew(Nil), nme.getClass_), Nil)