Fix problem with names encoding in scalap.
The refactoring performed in 020053c321 made use of naming encoding
more consisted but introduced a regression in scalap. The problem
is that the old encoder that scalap had didn't escape any characters
that were not included in its opcode list.

`NameTransformer` performs full encoding so it also encodes dots
that are being used as separators for packages. Therefore, in order
to retain the old behaviour we need to split the name by dots before
feeding each fragment to `NameTransformer`.

Review by @paulp.