SI-6407 Default to UTF-8 for text encoding
Traditionally there were two ways of getting a “default” text encoding:
 - Use UTF-8
 - Use the underlying platform's encoding

We shouldn't invent an additional, third way to do that.

Instead we use UTF-8 by default, and encourage people to be explicit.
Most of the code written out there is never designed with arbitrary
encoding changes via `-Dfile.encoding` or other means in mind, and only
causes pain, suffering and a larger testing surface for developers.

Deprecates the `Codec.ISO8859` constant, because the number is NOT
optional. "ISO-8859" is like "UTF" ... it's not meaningful and doesn't
denote a specific encoding.

Also fixes code that used `toLowerCase` or `toUpperCase` without a
locale. This is dangerous as code and scalac itself will behave
differently under some locales, namely azeri, lithuanian and turkish.