Moved test from neg to pending/pos
With nested classes enabled this test started to fail.  It turns out the error is only linked to nested objects and my suspicion is it's because of the same problems that caused lazy vals to fail with value classes. So the test should pass. I have re-opened the ticket and assigned it to @hubertp.