Fix for SI-5604, PACKAGE OBJECTS.
A major assist from retronym was all I needed.
Given what's in this commit it's amazing that package
objects haven't proven to be even buggier.  The major
culprit was mkAttributedSelect, which failed in its
effort to identify when the qualifier was a package object.

If these conditions held when typing an identifier:

 - its symbol is imported from a package object
 - the package object is not being compiled in the same run
 - the package object inherits the symbol from some parent

Then the symbol would not be recognized as requiring
`package` qualification, leading to all the package
object related cursing we've come to know and love.
This should herald a new era of somewhat fewer mysterious
package object behaviors.